+++ CHANGELOG.rst
@@ -1,6 +1,34 @@
 Changelog
 ---------
 
+0.38.0 (2018-06-10)
++++++++++++++++++++
+
+Features:
+
+- [apispec.core]: *Backwards-incompatible*: Rename ``apispec.utils.validate_swagger``
+  to ``apispec.utils.validate_spec`` and
+  ``apispec.exceptions.SwaggerError`` to ``apispec.exceptions.OpenAPIError``. 
+  Using ``validate_swagger`` will raise a ``DeprecationWarning`` (:issue:`224`).
+- [apispec.core]: ``apispec.utils.validate_spec`` no longer relies on
+  the ``check_api`` NPM module. ``prance`` and
+  ``openapi-spec-validator`` are required for validation, and can be
+  installed using ``pip install 'apispec[validation]'`` (:issue:`224`).
+- [apispec.ext.marshmallow]: Deep update components instead of
+  overwriting components for OpenAPI 3 (:issue:`222`). Thanks
+  :user:`Guoli-Lyu`.
+
+Bug fixes:
+
+- [apispec.ext.marshmallow]: Fix description for parameters in OpenAPI 3
+  (:issue:`223`). Thanks again :user:`Guoli-Lyu`.
+
+Other changes:
+
+- Drop official support for Python 3.4. Only Python 2.7 and >=3.5 are
+  supported.
+
+
 0.37.1 (2018-05-28)
 +++++++++++++++++++